Remove sync examples and clean up code
