+++ services/src/main/java/io/scalecube/services/ServiceDefinition.java
@@ -3,6 +3,7 @@ package io.scalecube.services;
 import java.lang.reflect.Method;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
+import java.util.Collections;
 import java.util.Map;
 
 
@@ -10,7 +11,6 @@ public class ServiceDefinition {
 
   private final Class<?> serviceInterface;
   private final String serviceName;
-
   private final Map<String, Method> methods;
 
   /**
@@ -18,28 +18,12 @@ public class ServiceDefinition {
    * 
    * @param serviceInterface the class of the service interface.
    * @param serviceName - the qualifier of the service.
-   * @param method - the method to invoke the service.
+   * @param methods - the methods to invoke the service.
    */
   public ServiceDefinition(Class<?> serviceInterface, String serviceName, Map<String, Method> methods) {
     this.serviceInterface = serviceInterface;
     this.serviceName = serviceName;
-    this.methods = methods;
-  }
-
-  private Type extractReturnType(Type type) {
-    if (type instanceof ParameterizedType) {
-      return ((ParameterizedType) type).getActualTypeArguments()[0];
-    } else {
-      return Object.class;
-    }
-  }
-
-  public Type returnType(String name) {
-    return methods.get(name).getReturnType();
-  }
-
-  public Type parametrizedType(String name) {
-    return extractReturnType(methods.get(name).getGenericReturnType());
+    this.methods = Collections.unmodifiableMap(methods);
   }
 
   public Class<?> serviceInterface() {
@@ -50,9 +34,12 @@ public class ServiceDefinition {
     return serviceName;
   }
 
-
   public Method method(String name) {
     return methods.get(name);
+  }
+
+  public Map<String, Method> methods() {
+    return methods;
   }
 
   @Override
@@ -61,10 +48,6 @@ public class ServiceDefinition {
         + ", serviceName=" + serviceName
         + ", methods=" + methods
         + "]";
-  }
-
-  public Map<String, Method> methods() {
-    return methods;
   }
 
 }
