+++ services/src/main/java/io/scalecube/services/annotations/AnnotationServiceProcessor.java
@@ -10,7 +10,6 @@ import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -39,11 +38,8 @@ public class AnnotationServiceProcessor implements ServiceProcessor {
           ServiceMethod methodAnnotation = method.getAnnotation(ServiceMethod.class);
           return resolveMethodName(method, methodAnnotation);
         }, Function.identity()));
-    
-    ServiceDefinition serviceDefinition = new ServiceDefinition(serviceInterface, serviceName,
-        Collections.unmodifiableMap(methods));
 
-    return serviceDefinition;
+    return new ServiceDefinition(serviceInterface, serviceName, Collections.unmodifiableMap(methods));
   }
 
   private String resolveServiceName(Class<?> serviceInterface, Service serviceAnnotation) {
